+++ frontend/src/app/services/ordinals/cat21.service.helper.ts
@@ -1,5 +1,11 @@
+import * as ecc from '@bitcoinerlab/secp256k1';
+import { base64, hex } from '@scure/base';
+import * as btc from '@scure/btc-signer';
 import ECPairFactory, { ECPairInterface, networks } from 'ecpair';
-const ecc = require('@bitcoinerlab/secp256k1');
+import { bytesToHex } from 'ordpool-parser';
+import { Observable } from 'rxjs';
+import { BitcoinNetworkType, signTransaction, SignTransactionResponse } from 'sats-connect';
+
 
 /**
  * Determines the minimum UTXO size based on the Bitcoin address type.
@@ -34,7 +40,6 @@ export function getMinimumUtxoSize(address: string): number {
   throw new Error('Unsupported address type');
 }
 
-
 /**
  * Creates a random SECP256k1 keypair via the ECC library
  */
@@ -46,6 +51,7 @@ export function createRandomPrivateKey(isMainnet: boolean): ECPairInterface {
 
 /**
  * Returns a hardcoded keypair
+ * This keypair should ne NEVER user for real transactions
  */
 export function getHardcodedPrivateKey(isMainnet: boolean) {
   const network = isMainnet ? networks.bitcoin : networks.testnet;
@@ -54,4 +60,102 @@ export function getHardcodedPrivateKey(isMainnet: boolean) {
 
   const ecPair = ECPairFactory(ecc);
   return ecPair.fromWIF(isMainnet ? mainnetDummyWIF : testnetDummyWIF, network);
+}
+
+/**
+ * Creates an input script for the Xverse wallet
+ * (the payment address for Xverse is always a P2SH-P2WPKH)
+ */
+export function createInputScriptForXverse(paymentPublicKey: Uint8Array, network: typeof btc.NETWORK) {
+  const p2wpkh = btc.p2wpkh(paymentPublicKey, network);
+  const p2sh = btc.p2sh(p2wpkh, network);
+  return {
+    script: p2sh.script,
+    redeemScript: p2sh.redeemScript,
+  };
+}
+
+/**
+ * Creates an input script for the Leather wallet
+ * (the payment address for Leather is always a P2SH-P2WPKH)
+ */
+export function createInputScriptForLeather(paymentPublicKey: Uint8Array, network: typeof btc.NETWORK) {
+  const p2wpkh = btc.p2wpkh(paymentPublicKey, network);
+  return {
+    script: p2wpkh.script,
+    redeemScript: undefined,
+  };
+}
+
+// see https://github.com/leather-wallet/extension/blob/8dbfefe8fcf5de687c2a137bce5eb2ff7a94b794/src/shared/rpc/methods/sign-psbt.ts#L49
+export interface LeatherSignPsbtRequestParams {
+  hex: string;
+  allowedSighash?: any[];
+  signAtIndex?: number | number[];
+  network?: 'mainnet' | 'testnet' | 'signet' | 'sbtcDevenv' | 'devnet'; // default is user's current network
+  account?: number; // default is user's current account
+  broadcast?: boolean; // default is false - finalize/broadcast tx
+}
+
+export interface LeatherPSBTBroadcastResponse {
+  jsonrpc: string;
+  id: string;
+  result: {
+    hex: string;
+  };
+}
+
+export async function signTransactionLeather(psbtBytes: Uint8Array, isMainnet: boolean): Promise<LeatherPSBTBroadcastResponse> {
+
+  const network = isMainnet ? 'mainnet' : 'testnet';
+  const psbtHex = bytesToHex(psbtBytes);
+
+  const signRequestParams: LeatherSignPsbtRequestParams = {
+    hex: psbtHex,
+    allowedSighash: [btc.SigHash.SINGLE_ANYONECANPAY],
+    signAtIndex: 0,
+    network,
+    broadcast: false // we will broadcast it via the Mempool API
+  };
+
+  // Sign the PSBT (and broadcast)
+  const result: LeatherPSBTBroadcastResponse = await (window as any).btc.request('signPsbt', signRequestParams);
+  return result;
+}
+
+export function signTransactionAndBroadcastXverse(psbtBytes: Uint8Array, paymentAddress: string, isMainnet: boolean): Observable<{ txId: string }> {
+
+  const networkType = isMainnet ? BitcoinNetworkType.Mainnet : BitcoinNetworkType.Testnet;
+  const psbtBase64 = base64.encode(psbtBytes);
+
+  return new Observable<{ txId: string }>((observer) => {
+
+    signTransaction({
+      payload: {
+        network: {
+          type: networkType
+        },
+        message: 'Sign Transaction (CAT-21 Mint)',
+        psbtBase64,
+        broadcast: true,
+        inputsToSign: [
+          {
+            address: paymentAddress,
+            signingIndexes: [0],
+            sigHash: btc.SigHash.SINGLE_ANYONECANPAY // 131
+          },
+        ],
+      },
+      onFinish: (response: SignTransactionResponse) => {
+
+        const txId = response.txId || '';
+
+        observer.next({ txId });
+        observer.complete();
+      },
+      onCancel: () => {
+        observer.error(new Error('Request was cancelled'));
+      }
+    });
+  });
 }
